created exercise card component and added valibot validation for later
